[inkling] Reorganize the model plugin into layers/model modules
Follow the kimi_k3 module layout: layers.py holds the transformer layer
classes (attention, router, shared experts, MoE/dense MLP), model.py holds
config construction, layer/block specs, the GPTModel subclass, and the
provider entry point. __init__.py lazily re-exports the two entry points.
Launch scripts point at miles_plugins.models.inkling.model.
