+++ scripts/models/inkling-975b.sh
@@ -38,5 +38,5 @@ MODEL_ARGS=(
     --qk-layernorm
 
     # Inkling model provider
-    --custom-model-provider-path miles_plugins.models.inkling.inkling.inkling_model_provider
+    --custom-model-provider-path miles_plugins.models.inkling.model.inkling_model_provider
 )
